backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
biopython==1.81
pysam==0.21.0
python-multipart==0.0.6
pydantic==2.4.2
numpy==1.24.3
anthropic==0.8.1 